Simplify forked executor
